requirements.txt
pillow
mss
supabase
utils
requests
python-dotenv
PySide6
